split the tests to the different files
